+++ tests/unit/container/test.c
@@ -22,25 +22,4 @@
  * SOFTWARE.
  */
 
-#include "writer/writer/test.h"
-
-extern cqueue_t *msg_queue;
-
-/*
- * O.K.
- */
-Test(writer_writer, test_writer_init0, .fini=test_stop_writer)
-{
-	cr_assert_eq(YALL_SUCCESS, writer_init(60));
-	cr_assert(msg_queue);
-}
-
-/*
- * Could not start thread and thus writer
- */
-Test(writer_writer, test_writer_init1)
-{
-	disable_pthread_create();
-	cr_assert_eq(YALL_CANT_CREATE_THREAD, writer_init(60));
-	enable_pthread_create();
-}
+#include "container/test.h"