+++ resources/build.sh
@@ -25,8 +25,8 @@ build-wrapper-linux-x86-64 --out-dir ${YALL_ROOT}/build/out/bw make -C ${YALL_RO
 make -C ${YALL_ROOT}/build test
 make -C ${YALL_ROOT}/build doc
 
-if [ "$TRAVIS_PULL_REQUEST" = "false" ] && [ "$BRANCH" = "develop" ]; then
-    SONAR_EXTRA_OPTIONS="-Dsonar.branch.target=develop -Dsonar.branch.name=${BRANCH}"
+if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
+    SONAR_EXTRA_OPTIONS="-Dsonar.branch.name=${BRANCH}"
 else
     SONAR_EXTRA_OPTIONS="-Dsonar.pullrequest.branch=${BRANCH} -Dsonar.pullrequest.key=${TRAVIS_PULL_REQUEST} -Dsonar.pullrequest.base=develop -Dsonar.pullrequest.provider=github -Dsonar.pullrequest.github.repository=${GITHUB_REPO}"
 fi