+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 * Thread loop timer (integer overflow)
+* Stack overflow when too many messages are printed
 ### Security
 
 ## [1.7.7]