+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+* Queue (non concurrent) mechanism
 ### Changed
 * Set queue as cqueue (concurrent queue)
 ### Deprecated